ISSUE-47193049: Branch commit. Restorepoint. Remote & local machines unit tests are working

+++ tasks_branches/branched_from_v1_1/ISSUE-47193049_StateMachines/Code/Main/ermeX.ComponentServices/RemoteComponent/IRemoteComponent.cs
@@ -6,6 +6,6 @@ namespace ermeX.ComponentServices.RemoteComponent
 	internal interface IRemoteComponent : IErmexComponent
 	{
 		void Join();
-		void Create(Guid componentId, IPAddress ipAddress, int port);
+		void Create(Guid componentId, IPAddress ipAddress, ushort port);
 	}
 }
